pyproject.toml
[project]
name = "oronder"
version = "1.0.0"
description = "Discord + FastAPI service for Oronder/Foundry integrations"
authors = [
    { name = "Tuna", email = "chris@oeleri.ch" }
]
requires-python = ">=3.13"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    "aiohttp==3.12.15",
    "brotlipy==0.7.0",
    "d20==1.1.2",
    "dateparser==1.2.2",
    "fastapi==0.117.1",
    "httpx==0.28.1",
    "Jinja2==3.1.6",
    "markdownify==1.2.0",
    "msgspec==0.19.0",
    "numpy==2.3.3",
    "Pillow==11.3.0",
    "psutil==7.1.0",
    "psycopg2-binary==2.9.10",
    "py-cord @ git+https://github.com/Pycord-Development/pycord",
    "pydantic==2.11.9",
    "python-socketio==5.13.0",
    "pytz==2025.2",
    "pytzdata==2020.1",
    "setuptools==80.9.0",
    "SQLAlchemy==2.0.43",
    "tabulate==0.9.0",
    "uvicorn==0.36.0",
]

[dependency-groups]
dev = [
    "ruff==0.13.1",
    "mypy==1.18.2",
    "types-tabulate==0.9.0.20241207"
]
